moved root directory .py files to project directory
